virt: Add missing seclabel's on migrations
Vdsm 4.3 enabled dynamic ownership.  It also ensures that appropriate
seclabel's are added to disks so that libvirt doesn't change their
permissions.

When a VM is migrated from a host running older Vdsm < 4.3, where
dynamic ownership is not enabled, the VM domain XML doesn't contain
the seclabel's.  When the VM is powered off in any way, libvirt
assumes defaults in such a case and changes the owners of the VM disk
images to root.  Once VM is attempted to start again, the disk images
are inaccessible, resulting in an error.

This patch installs a libvirt hook for both host and file migrations
that adds the missing seclabel's.

Change-Id: Ib77888fc6824a5a49fe4b66c0230a8f989090437
Bug-Url: https://bugzilla.redhat.com/1666795
Backport-To: 4.3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>