formatconverter: Convert file based domain to v5
Implement conversion for file based domains. In this flow we have 3
steps:

1. Converting volumes metadata to V5, keeping V4 keys
2. Convert domain metadata to V5
3. Remove V4 keys from volumes metadata.

Failure in step 1 or step 2 will fail the conversion, leaving the
domain in the current version.

Failure in the 3rd step is not fatal. The volumes metadata will be
updated on the next update to the volume.

Since converting file and block domain requires different
implementation, new conversion interface was added:

- convert_volumes_metadata(target_version)
  Perform step 1, adding metadata added in target_version to volumes
  metadata, keeping current version metadata.

- convert_metadata(target_version)
  Perform step 2, converting domain metadata to target_version.

- finalize_volumes_metadata()
  Perform step 3, removing keys not in target_version from volumes
  metadata.

Converting block storage domain is not implemented yet, since it
requires different approach. Converting block storage domain will raise
NotImplementedError.

Change-Id: I49f656a7ef5c7a695e07d56c0d69a4f59941c541
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>