storage: metadata slot offset supporting v4 and v5
Block volume metadata are stored in the first 1 MB in metadata LV
in version 4, while in version 5 this area is zeroed and metadata
is stored in area 1 - 17 MB. Therefore we need additional offset
in version 5 when computing location of volume metadata.

This chnage adds function which returns correct metadata offset
for provided version and this offset is used when computing
final metadata offset when reading or writing volume metadata.

Change-Id: I5f42028f51c3600d943486a4952521dc8d0f1e18
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>