storage: rename offset to slot
Block storage volume stores metadata slot in lvm tag. This slot
number variable is named as "offs" and method which reads it is
named "getMetaOffset". This is missleading as this is actually
metadata slot number and offset is computed from slot number.

Rename method for reading metadata slot number and local variable
to reflect actual meaning what the method does or variable stores.

Change-Id: I475f19be2cbd7dd2ab9c97c6e110231610611c5e
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>