volume: Make _getMetaVolumePath public
We need this to verify that volume metadata is written correctly to
storage.

Since this method is not implemented in BlockVolume, and does
not make sense, remove it from Volume interface. This requires
silencing pylint check, since it cannot tell that the instance manifest
is a FileVolumeManifest.

Add a test for verifying writing volume metadata to storage, using the
new public method. The test will be more interesting when we add support
for setting additional metadata.

Change-Id: Ifa64a3d08ddbc231355b97ce9118045aed3c0f6f
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>