makefile: Added 'fix-python-shebangs' dist hook
This patch introduces a dist hook, that will replace all occurences of
'#!/usr/bin/python[23]?' shebangs with ones that use
'VDSM_TARGET_PY_MAJOR_VERSION' variable. The hook doesn't modify your
tree - it works in the dist directory, which is then archived.

Warning! Currently a lot of our are scripts are *not* yet
py3-compatible - this is a build-oriented change only.

Change-Id: I807e8982fbca0027f138d9c8a9732625c2cc159a
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>