volumemetadata: Don't convert metadata to string
Since the initial commit in 2011, Volume.setMetadata() was wrongly
converting metadata values to strings. This issue was hidden since the
value was stored in a dictionary, and immediately formatted and written
to storage.

Now we use VolumeMetadata, supporting both old code using SIZE in 512
bytes blocks, and new code using CAP in bytes, so we must work with
integer values only when setting size.

Fix Volume.setMetaParam() to store the given values without changing the
type.  The key will be validated and rejected if the key is not a string
and one of the allowed keys. The value will be stored as is an formatted
to string when formatting metadata.

Additionally, fix old wrong code in formatconverter, setting SIZE as
string.

Change-Id: I5a03d799e4e44df917aea0937694b6a1748b6c3f
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>