caps: Report supported domain versions
When mixing hosts with different vdsm versions, we may have hosts that
support the new storage format v5 (version > 4.30.9) and hosts
supporting only storage format v4 (version <= 4.30.9).

Engine needs to know the supported domain version in these flows:

- Creating the first domain in a DC using cluster level 4.3. If all
  hosts in the DC support domain version 5, create a v5 master domain.

- Attaching a host to a DC using v5. If the host does not support V5,
  engine should prevent the operation.

Add a "domain_versions" key to Host.getCapabilities reporting the
supported storage domain versions.

When working with hosts not reporting this feature, engine must assume
that a host does not support storage domain v5.

Supporting v5 does not ensure that a host supports 4k. We will need to
report this capability in a similar way when we actually support 4k.

This information will not be needed in cluster version 4.4, when all
hosts will support v5 and 4k storage.

Change-Id: I7ea0360d83fa47983c5985844de9243e4de6774d
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>