CI tests: Add net functional tests
Adding functional tests to the CI as a custom stage. This
'check-network' stage will not run automatically, but
will need to be manually trigerred through jenkins.

The functional tests that are added are tested with legacy
switch (linux-bridge).

There is a way to run the manual tests with the Jenkins GUI
through the following link [1], but a few fields should be filled:
 - STD_CI_CLONE_URL - hardcoded 'git://gerrit.ovirt.org/vdsm'
 - STD_CI_STAGE - hardcoded 'check-network'
 - STD_CI_REFSPEC - few steps are required:
	1) Make a request through the gerrit API to [2], where patch
	   number is the only variable to replace by the number of
	   the patch you want to run the tests on.
	2) Copy ['revisions'][current_revision]['ref'] value.

[1] https://jenkins.ovirt.org/job/standard-manual-runner/build
[2] http://gerrit.ovirt.org./changes/?q=patch_number&o=CURRENT_REVISION

Change-Id: Ie5b741ee710d70b62aa3dab154fc55b03b341d13
Signed-off-by: Bell Levin <blevin@redhat.com>