virt: Refactoring in VM._setTicketForGraphicDev
This is a follow-up patch to https://gerrit.ovirt.org/#/c/97381/

A discussion has started whether API of VM.setTicket should be changed
to accomodate datatypes other than strings in params map. At this point
the API uses a simple string->string map. Changing it is beyond the
scope of the original enhancement request.

This patch only introduces some refactoring.

Change-Id: If7ee8ff04188ccebf59704603720e23291f41962
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>