localfssd_test: add tests for creating volumes with initial size
Add a test for creating raw-sparse and cow-sparse volume with initial size
which is not allowed.

Also add a test for creating raw-preallocated volume with initial size
which is not allowed at this stage but will be implemented in the
following patches that will remove the xfail mark.

This patch is part of a series of patches that add the option to create
a raw preallocated volumes on file-based storage domain with an initial size.
This functionality is needed in order to keep the file volume format
when cloning it from a template and migrating it while the VM is down.

Change-Id: I93fc0583dd9572815f7b2619630739e40d749ea3
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>