fileVolume: split _create() to raw and cow volume creation helpers
This patch splits the implementation of _create() to two helpers:
_create_raw_volume() - will handle the creation of volume with RAW format.
_create_cow_volume() - will handle the creation of volume with COW format.

Each helper contains only the relevant code parts for the specific
volume creation.

This patch is part of a series of patches that add the option to create
a raw preallocated volumes on file-based storage domain with an initial size.
This functionality is needed in order to keep the file volume format
when cloning it from a template and migrating it while the VM is down.

Change-Id: Iaa6fb857d16b29f731b56d85bd931b8ecf7727b9
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>