fileVolume: extract file allocation to a helper method
Exract file allocation using fallocate to a helper method
to avoid code duplication.

This patch is part of a series of patches that add the option to create
a raw preallocated volumes on file-based storage domain with an initial size.
This functionality is needed in order to keep the file volume format
when cloning it from a template and migrating it while the VM is down.

Change-Id: I638b929aef1f46bce309ac4d9e4965384ff7c45f
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>