fileVolume: change _create() helpers to pep8 code style
This patch align the code in _create_cow_volume() and
in _create_raw_volume() to pep8 code style.

This patch is part of a series of patches that add the option to create
a raw preallocated volumes on file-based storage domain with an initial size.
This functionality is needed in order to keep the file volume format
when cloning it from a template and migrating it while the VM is down.

Change-Id: Ibb3541d57d40378b401a18ea6b8782289d110ad9
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>