virt: Wrong attribute tested for VNC security
When a VM start is requested, vdsm checks if VNC configuration is
secure. It either has to have a password or have SASL enabled.

A wrong attribute of the graphics tags has been tested. Instead of
'passwd', the attribute 'passwdValidTo' is the one that libvirt uses to
enable/disable password authentication.

Change-Id: Ib646f31a69be6a128668468137eedd6c1a2e7b58
Bug-Url: https://bugzilla.redhat.com/1695567
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>