fileVolume: use bytes instead of blocks in _create() helpers
This patch convert the sizes units from block to bytes in
_create_raw_volume() and _create_cow_volume() helpers.

This patch is part of a series of patches that add the option to create
a raw preallocated volumes on file-based storage domain with an initial size.
This functionality is needed in order to keep the file volume format
when cloning it from a template and migrating it while the VM is down.

Change-Id: I1c2c7b490f0c9548c3cd262be9f4efd28ed97f13
Bug-Url: https://bugzilla.redhat.com/1571285
Bug-Url: https://bugzilla.redhat.com/1403183
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>