virt: Remove support for legacy migrations
Legacy ("downtime", non-schedule) migrations are going to be removed
from 4.3 cluster levels.  That means we needn't support them in 4.4
anymore and the corresponding code can be removed.

We still can't exclude that a legacy migration is invoked by Engine.
That shouldn't be a regular flow so we use a fallback migration
schedule in such a case rather than keeping the old code around.

Change-Id: Ib789c97a077e2c5824091d4a24930bc5dea68339
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>