py3: Drop 'jsonrpcclient' from py3k blacklist
This patch fixes 3 issues found by 'pylint-py3k'.

Replacing 'self._responses.values()' with explicit conversion
to a list - 'JsonRpcClient.callMethod' uses indexing to access
the first element and 'JsonRpcCall.callback' is already wrapping
single responses in a list.

Replacing 'self._responses.keys()' with 'six.iterkeys(...)' is safe
because it's only used in an iteration-context
in 'JsonRpcClient.call_cb'.

Call to 'map' has been replaced with a simple 'for' loop because we
don't care about the value, just the side effects.

Change-Id: Ie53e6882607eb5cf82e46788cb3e434cbb87e449
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>