py3: Drop 'stomp' from py3k blacklist
Two changes are introduced to make pylint-py3k happy.

First one simply ignores the 'round-builtin' warning - the way rounding
works changed between py2 and py3 ("away from 0" vs "Banker's
Rounding"), but in this case, we're not so precision-dependent to care.

Second change uses 'six.text_type' instead of 'unicode' to filter-out
the type of the variable, thus removing reference to 'unicode'.
This is a semantic change only.

Change-Id: I121ce005e875f1205705be9c4a918cc902621502
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>