virt: Reset post-copy flag at the beginning of a migration
Vdsm assumed that once post-copy migration is started the VM is gone
and will never be migrated again.  The assumption appeared to be
wrong: We have experienced a case when a VM is resumed on the source
after being killed on the destination in post-copy (which may or may
not be a correct behavior of libvirt/QEMU); and additionally at least
in theory the migration may not enter post-copy after a successful
post-copy switching request.

Let's fix that by resetting Vm._post_copy at the beginning of each
migration.

Change-Id: I0f09a880e3dc69cb5f521c7861cd5560115aa17d
Bug-Url: https://bugzilla.redhat.com/1688159
Backport-To: 4.3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>