virt: Don't finish incoming migration when switching to post-copy
When a migration switches to post-copy, it starts running on the
destination and Vdsm starts performing normal actions on it such as
writing metadata or running periodic jobs.  However, as long as the
libvirt post-copy migration job is running, libvirt may reject running
other jobs on the VM.

For instance, metadata writing may block in libvirt until migration
finishes or the request times out, whatever happens first.  In the
latter case metadata write attempt raises a libvirt timeout error,
resulting in interruption of VM start and destroying the VM, with
fatal consequences for the post-copy migration.

Let's be on the safe side with libvirt operations and set
Vm._incoming_migration_vm_running only after the whole migration is
completed.

Change-Id: Iff75669ed03b60a71315d0cf9bfb0aa81d0d6a5d
Bug-Url: https://bugzilla.redhat.com/1688159
Backport-To: 4.3
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>