py3: net tests: error when sort() is used on tuple
The tests pass as expect via the the client, but an error is
thrown using the lib. The function _canonicalize_bond_slaves
expects a list, but a tuple is sent instead.

The reason this test passes via the service, is because the vdsm
client sends the setupNetworks info as a json, when decoded on
the service side, it is decoded as a list, and not as the original
tuple value.

Change-Id: Ie6b1664ccb6875d9af490f9b6d35db1768c334cf
Signed-off-by: Bell Levin <blevin@redhat.com>