net: py3: Setup vlanned nets first to fix QoS
When setting up (in a single transaction) a non vlanned
network, and a vlanned network on the same nic with QoS
configuration, the non-vlanned network's QoS configuration
is getting overriden by the vlanned QoS config.
When going the other way around, first initializing the vlanned
and then the non-vlanned, everything works as expected.

This is reproducible only on python3, because the way a dict
is ordered has changed between py2 and py3. One of the network
tests currently (coincidentally) pops the non-vlanned network
first on py3.

The current workaround is to arrange the networks in a way where
the vlanned are treated first.

Reproduced by the following test:
test_add_two_networks_with_qos_on_shared_nic

Change-Id: Iffe74f979c729584a6788c96e03cc2738af742b5
Signed-off-by: Bell Levin <blevin@redhat.com>