configure.ac: Added '--with-target-python' argument
Added a '--with-target-python' argument that allows to define which
interpreter version we're targeting. Example usage:

    ./autogen.sh --with-target-python=python3

Both names like 'python2' or 'python3' and full names like 'python3.6'
are allowed. In the former case, an attempt will be made to resolve
the ambiguous name to full name in the context of the environment.

This option changes only the version we're targeting. If other
interpreter versions are found in the environment, they will be
supported - packages will be required for them, test will run for them.

Change-Id: If5c102542420c3e970efee8dd89f49e13d7252ca
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>