configure.ac: Added '--with-only-python' option
Added '--with-only-python' option to 'autogen.sh'. Example usage:

    ./autogen.sh --with-only-python=python3

Both names like 'python2' or 'python3' and full names like 'python3.6'
are allowed. In the former case, an attempt will be made to resolve the
ambiguous name to full name in the context of the environment.

This flag causes the build system to completely ignore other interpreter
versions found in the OS - no packages will be required for them and no
tests will be run using these.

Change-Id: I9917643205d08e0ccf05c419d357fdb9dbcb4151
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>