m4: Allow 'pythonMAJOR' interpreter name form
This change loosens the restriction on the interpreter name form that
'VDSM_CHECK_PY_VERSION' macro accepts. Now, when passed a name like
'python3', an attempt will be made to resolve it to a full name like
'python3.6'.

The algorithm for resolution uses symbolic link following
until a satisfactory name is found, or an end of symbolic links chain
reached. In the latter case, an error will be raised.

Change-Id: I2ee6ff2c55b3a8ad27b803cf567182a65dec6157
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>