configure.ac: Limit supported interpreter versions
Limiting supported py3 interpreter versions to the default one available
in the OS for now (the one 'python3' points to).

Many python packages, which are test dependencies, are listed only
in 'automation/*packages*' files. To support py3 versions different,
than the default one in the OS, we need to add them also to 'tox.ini'.

Since currently, there is no 'selinux' python bindings package in PyPI
repository, tests that import (directly, or indirectly) 'selinux'
package will have to be disabled for non-target interpreter versions.

This will be fixed in next patches by adding dependencies to 'tox.ini'
and marking some tests as 'target-only'.

Change-Id: Iaccfd083f975aff268fcfe611ec78ea707fd0734
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>