backup: add scratch disks creation for start_backup
When a backup starts, a scratch disk should be created for each disk
that the backup includes.

In order to create this scratch disks, we have two options -
1) Let Libvirt create/remove the scratch disks for us.
2) Create and remove the scratch disks in a custom way.

The problem in the first option is that libvirt will create a scratch
disk on the same path of the original volume.
We cannot allow it since we don't want to create and maintain anything
under the shared storage volume paths.

So this patch implements the second option, when a backup starts,
all the scratch disks will be added using the transientdisk API.

When the backup stops, the scratch disks will be removed.

Example for the backup XML that includes specific VM disks-
<domainbackup mode='pull' id='0'>
  <server transport='unix' socket='/run/vdsm/backup-id'/>
  <disks>
    <disk name='vda' type='file'>
      <seclabel model="dac" relabel="no" type="none"/>
      <scratch file='/path/to/scratch/disk.qcow2'/>
    </disk>
    <disk name='sda' type='file'>
      <seclabel model="dac" relabel="no" type="none"/>
      <scratch dev='path/to/scratch/disk.qcow2'/>
    </disk>
  </disks>
</domainbackup>

Scratch disks created and removed using HSM transient disk APIs.

Change-Id: I641bc4d3a7c38c52dce5e4a189d12d019d37e39a
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>