net tests: sort arp_ip_target before assert
The arp_ip_target test fails intermittently on fedora alone.
The reason for the fail is the assertion at the end of the
setupNetworks (and bond asserts) - in Fedora.

While fetching the kernel config using sysfs command, (which
is the "culprit" here, returning different order every run),
the arp_ip_target value is in different order.

For example, two arp_ip_targets strings returned from the
same test on different runs:
"10.1.3.1,10.1.2.1"
      ^        ^
"10.1.2.1,10.1.3.1"
      ^        ^

I was not able to find any refernce to real life problem
involving arp_ip_target changing it's order, thus the problem
is in the asserts alone.

The proposed fix is to normalize both running and kernel configs,
by sorting the arp_ip_target before asserting them.

Change-Id: Ied74df8bbe289c39529dd0e66062dd847bf6af93
Signed-off-by: Bell Levin <blevin@redhat.com>