vm_libvirt_hook: Rename libvirt/ to libvirthook/
Having a package named libvirt in the virt package means that no module
can import the real libvirt module, which is pretty sad.

Turns out that tests/virt/vm_libvirt_hook.py test was broken, trying to
import the hook from vdsm.virt. The issue was hidden since the test
module was missing a _test suffix, so it was not never collected.

Fixing the import reveal that the hook is broken on python 3, or maybe
the way we run the hook in the test is incorrect. The tests are skipped
on python 3 until someone has more time to look at this.

Change-Id: I681c24820843dc6920038c454e0c816d7151373d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>