tests: init lockspace in clusterlock tests
Sanlock lockspace wasn't initialized in clusterlock tests. This is
a bug in fake sanlok, which doesn't check if the lockspace was
initialized before it's used. This will we fixed in subsequent patches.
This patch add correct lockspace initialization into the tests.

To reduce the code, fixture for creating and initalizing lock is created
and used in the existing tests. With proper lock initialization,
dedicated fake_sanlock fixture is not needed any more and is removed in
this patch.

Change-Id: I6c66c2841d3eb6a4fae155f7675e46654532b256
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>