volume: remove size override when creating a volume with parent
When creating a volume with a parent, for example when moving a disk with
a snapshot to a different storage domain, the newly created volume size
that send by the engine overrides.

size = volParent.getSize()

This was introduced in commit 8a0236a2fdf4e81f9b73e9279606053797e14753
There is no info explaining why vdsm need to override the size sent by engine
silently and use the parent size.
Maybe this was needed in the past to work around some engine bug or issue in
another vdsm flow.

This behavior caused a corruption of the VM metadata when a disk with
a snapshot is extended and moved to a different storage domain.

This patch removes the size override.

Change-Id: Ie7652a3f0c7bee907c7aeae0e1b8bcfcd11247af
Bug-Url: https://bugzilla.redhat.com/1707932
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>