py3: Fix rollback exception not iterable anymore
test_rollback_target_configuration_failed_connectivity_check is
not passing on py3 due to exceptions not being iterable on py3
anymore.

A fix for this issue is to refactor the exception to be an
object and contain the values as attributes.

- Moved to "new style" python because of a pylint issue
menifesting when applying the following simple fix:
x, y, z = err.args
instead of:
x, y, z = err
Three pylint errors were thrown due to this issue [1], failing
the check-patch.

[1] https://github.com/PyCQA/pylint/issues/2621

Change-Id: Ib41a20a332a78e6de64b95762c36a2acea3c87c3
Signed-off-by: Bell Levin <blevin@redhat.com>