net tests: split post setup hook
When setupNetwork fails, the post_setup_hook does not run,
thus the updating of the runnning config and kernel config are
not preformed. This results in the value netinfo not being
initialized.

When an assert is required to be preformed after setupNetworks
(e.g. assertNoNetworkExists), the assert fails due to an AttributeError.
self.netinfo.network - which is the asserted value, is not initialized,
because the initialization is part of the setupNetworks post hook.

The update of the configs should be ran regardless of the success
status of setupNetworks, because the value are needed later on,
even if the setup failed.

Change-Id: If2883e32c263ee5b0b397db9172e36c08d3f4761
Signed-off-by: Bell Levin <blevin@redhat.com>