storage: add getters for alignment and block size
Add methods for accessing storage domain alignment and block size.

Values are loaded from storage domain metadata in the constructor and
stored as both of them cannot be changed during domain life cycle.
Their values are also validated and we raise an exceptions when
the value is invalid.

These function will be used later on for passing these values into
sanlock methods which require this information.

Change-Id: I0bc1a63b7827fdbc8d5f544c8424339ceaabe174
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>