virt: tests: Remove some legacy tests
Legacy (non-Engine-XML) device configuration is going to be removed
soon.  We need to convert unit tests using legacy configuration to
Engine XML.  As a part of this change, we need to switch to Engine XML
in fake objects.  But before we can do that, we must make sure unit
tests using legacy configuration won't break.

Let's start with the easiest step -- removing some tests that check
only legacy paths.

Change-Id: I50700586e63909d2df7fc1fc6baa00eb67e5bf16
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>