storage: add 4k support into xlease
This patch adds 4k sanlock support for xleases - if available sanlock
module supports 4k API, alignment and block size are passed to sanlock
calls. The call is handled by sanlock module wrapper which does actual
check if the 4k API is supported or not.

Please note that this patch only adds support for calling sanlock with
alignment and sector size arguments, but doesn't prepare this module for
4k storage. I.e. after integrating this patch, using xleases on 4k
storage or using alignment other than 1M is still NOT possible.

Change-Id: I03af36defe894141bcd6d4f9a6b03e263c8f99b2
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>