storage: add 4k support to fileVolume
This patch adds 4k sanlock support for fileVolume - if available sanlock
module supports 4k API, alignment and block size are passed to sanlock
calls. The call is handled by sanlock module wrapper which does actual
check if the 4k API is supported or not.

As the affected method newVolumeLease() isn't tested, this patch also
adds a test for this method and related changes in test infrastructure:
- monkeypatches sanlock with fakesanlock in fileVolume module
- adds missing getMetadata() method into FakeSD class

Change-Id: Ib389565f84dbfe191741737ba40f3480a1565857
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>