volume: Repair volume metadata capacity
volume.updateInvalidatedSize() repaired capacity stored in volume
metadata only when the capacity was zero. However, there can be more
sources of volume size inconsistencies so it would be convenient to use
this method in these cases as well.

This patch amends this method to:
- repair volume metadata capacity if it is smaller than the virtual size
  reported by qemu-img info, which covert both cases.
- skip shared volumes, which are immutable and must not be modified.

This patch also adds skipping of one merge test as monkeypatching done
in this test is too broad and cause unrelated failures. The test will
be fixed in followup patch.

Modifications for 4.2:
- Check and set SIZE in blocks instead of CAP in bytes
- Replace pytest.mark.xfail with testValidation.xfail

Change-Id: I4f4f91d1e2fcbc3a4a72162fb8dbc1b24b73576a
Bug-Url: https://bugzilla.redhat.com/1707934
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>