net, sriov: Use netlink for update_numvfs check
Upon creation of the SR-IOV virtual functions
the vdsm was waiting only for the udev events to
settle down. This resulted in the wrong caps being
reported in case some of the events appeared after
the 'settle' function return.

Add a monitor that waits for the VF/s up event,
confirming that the operation succeeded. This
approach assures the end result (iface created and up)
and is not dependent on timing ("slower" drivers).

The monitor is present with timeout just in case the operation
takes longer than expected to fail with exception.

Change-Id: Ib5d615fd1d31b157300e1b9cb330247d836de753
Bug-Url: https://bugzilla.redhat.com/1695037
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit b53c9a53d5de38b005fc199353f05478813b5688)