volume: Validate volume size when creating snapshot
In the past we used to ignore the requested size when creating a volume
with a parent volume UUID. Now that we actually use the size, we need to
validate it.

Verify that volume size is not smaller than the parent size. Creating
such volume will cause image corruption when qemu try to access offset
beyond the volume size.

The tests are separated to allow backporting the fix to 4.2.

Change-Id: Ic2cc8afcf61dfcee08f9ca7d8fbc30cb1c46e61c
Bug-Url: https://bugzilla.redhat.com/1707932
Signed-off-by: Nir Soffer <nsoffer@redhat.com>