net: Remove custom bond options
Custom bond options were used in the past to retain info about
ovs [1]. All usages have been removed since, thus rendered useless.

One piece of code concerning the custom bond options was kept,
namely the popping of the custom option in the kernelconfig.py.
This is due to the custom option being persisted, thus opening
the possibility to this option existing in an old running config.
Considering parsing of the custom bond is removed, an old
configuration might break the parsing during a net-restore.

[1] https://gerrit.ovirt.org/#/c/70033/

Change-Id: Iad37cd97e269f5c72bb7aeb87ad35ba45c46667b
Signed-off-by: Bell Levin <blevin@redhat.com>