virt: Changed errors to warnings for deployments
When deploying a Hosted Engine on a Host the
ovirt-ha-broker and ovirt-ha-agent services are
stopped and then restarted. During that time period
the vdsm attempts to continue to send get_stats
messages which leads to a failure because neither
the client nor the server are up to service the
requests. This causes exception errors within the
journal log as well as on the vdsmd status. This
caused user complaints because they believe it is
a malfunction.

The solution is to change the errors and exceptions
to warnings so that they do not appear as true errors,
but still provide diagnostics when there truely is a
communication issue. Setting timeouts does not help
because neither the broker nor agent are running when
the problem occurs.

Change-Id: I83d2a2069163926bca7c7b0b85dc36e4b456e73e
Bug-Url: https://bugzilla.redhat.com/1569593
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>