ci: Add '*environment.yaml' files
This patch adds two files, and a couple of symlinks that point to them,
that define environment variables for our CI stages/sub-stages.
Currently only one variable is being defined - ${CI_PYTHON_VERSION}.
This variable defines the name of the interpreter we will use when
configuring the build tree and running 'tests/profile' script.

Change-Id: Ia90d79cf599987af995486dcd3602a079f7b7616
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>