ci: Use ${CI_PYTHON_VERSION} to run 'tests/profile'
To cover cases where an OS doesn't have Python 2 interpreter
at all, we need to specify explicitly which Python version to use when
running 'tests/profile'. Unfortunately this script is used even before
anything is built - there is no way to generate it from a '.py.in'
file to change the shebang line. This patch uses '${CI_PYTHON_VERSION}'
variable to use appropriate interpreter version to run it.

Change-Id: Id108cf10d012a11cd0e4752145d967b86a062c9d
Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>