storage: add 4k support for clusterlock
This patch adds 4k sanlock support for clusterlock module - if available
sanlock module supports 4k API, alignment and block size are passed to
sanlock calls. The call is handled by sanlock module wrapper which does
actual check if the 4k API is supported or not.

When the domain is created, lock is created after validation of
alignment and block size to make sure we initialize it with correct
values.

Change-Id: I6311f14ae1e6c35686140c570f2a2a1255c1191a
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>