upgrade: Skip volumes with cleared metadata
vdsm version before 4.20.34-1 (ovirt 4.2.5) had a very old bug when
volume metadata was cleared by replacing the metadata with:

NONE=#####################################################...

after we failed to remove a LV. The result is a volume with invalid
metadata. Trying to access the volume metadata will fail with this
error:

    MetaDataKeyNotFoundError: Meta Data key not found error: ("Missing
    metadata key: 'SIZE': found: {'NONE': '###################...

This bug was fixed in block storage in commit 7bbb4b867249 (blockVolume:
Remove metadata only if the LV successfully removed). But storage
domains used with older versions may have such partly-deleted volumes.

When converting storage domain to V5, reading the metadata fail, and the
entire conversion is aborted. The next SPM trying to upgrade this domain
will fail in the same way, bringing the entire data center down.

To fix this issue, we detect now cleared metadata when parsing volume
metadata, and raise a new MetadataCleared exception. The code converting
volume metadata was updated to skip these volumes since we don't have
any metadata to convert.

With this change converting to V5 will succeed, leaving the
partly-deleted volumes on storage. We may need to consider deleting
these volumes in a future version.

A new warning is logged when detecting volumes with this issue:

    WARN Skipping volume 25869545-0dca-464e-91f1-6b7b6e332352 with
    cleared metadata, volume is partly deleted

Finding such volume is very unlikely, so the simplest solution of one
log per volume was chosen.

Change-Id: I789f358460cfa3fe32097e958a8261e75a5f9621
Relates-To: https://bugzilla.redhat.com/1574631
Bug-Url: https://bugzilla.redhat.com/1714154
Signed-off-by: Nir Soffer <nsoffer@redhat.com>