lvm: Log LVM warnings for successful commands
LVM has a known issue, when command that should not change the VG detect
inconsistency in the VG metadata, and try the fix the metadata. This can
happen when LVM command on the SPM changes VG metadata while LVM command
is running on another host. This can lead to VG metadata corruption.

Turns out that LVM logs this warning in this case:

    WARNING: Inconsistent metadata found for VG xxx-yyy-zzz - updating
    to use version 42

But this important warning is logged in debug level which is disabled by
default.

Change lvm commands to log a warning when LVM commands succeeds, but
stderr is non-empty. This does not fix the bug, but will help to
diagnose the issue.

An unwanted result of this change is filling the logs with this
unhelpful warning:

    WARNING: This metadata update is NOT backed up

Change-Id: Id668415ebea9399496d342a35011e4e8ea8ad565
Relates-To: https://bugzilla.redhat.com/1553133
Bug-Url: https://bugzilla.redhat.com/1715026
Signed-off-by: Nir Soffer <nsoffer@redhat.com>