spec: Require sanlock 3.7.3 providing 4k support
This version is available in Fedora 29 and 30, and on CentOS, in the
testing repo enabled by ovirt-release-master package.

On Fedora 28 the package is available in updates-testing repo. The repo
should be enabled manually by developers if they cannot use Fedora 29
yet.

In all cases requires only sanlock-python, since it requires sanlock.

Change-Id: Iee1b6709b50888dec7dc765709a790f22947bb39
Signed-off-by: Nir Soffer <nsoffer@redhat.com>