tests: check underlying sector size in fake sanlock
This is the last of the patches which add alignment and sector size
validation to the sanlock. This patch adds check that sanlock is
called with same sector size as underlying storage.

As fake sanlock keeps everything in memory and doesn't write anything
to disk, fake sanlock instance variable is added which represents
underlying sector size and the checks are done against this value.

The only exception when this check is skipped is during write_resource()
call, as real sanlock doesn't fail even when called with wrong sector
size.

Change-Id: Ic2e13b25b1015907fa17665ad19920538b14a962
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>