py3: Add storage tests for python 3.7
Add storage-py37 env, and fedora 29 build on travis that will run this
env.

More work is needed to add py37 envs for lib/virt/network. Until all the
teams add py37 test env, add simple test-py37 target that run only the
storage tests. We will modify this gradually until all teams can run the
tests with python 3.7.

The tests reveal a regression[1] in python 3.7 - when invoking
makedirs("a/b", mode), mode does not affect the mode of new intermediate
directories, so directory "a" mode will not use the specified mode. A
new test added to verify mode of intermediate directories, and marked as
xfail on python 3.7.

[1] https://docs.python.org/3/library/os.html#os.makedirs

Change-Id: I301bf97fced6d354619dd6d24f1d90b989595bae
Signed-off-by: Nir Soffer <nsoffer@redhat.com>