spec: Allow making vdsm.spec without running configure
Currently, vdsm.spec is generated from vdsm.spec.in using values
produced by configure.  That means vdsm.spec must be generated in the
same environment as the target build environment.  This is not always
desirable, i.e. it should be possible to build a SRPMS in a different
environment or build dependencies should be available before building
the package (chicken or egg problem).  These changes are especially
needed for current automated package building infrastructures.

This patch replaces the current configure variables with default
values.  While this duplicates some of the values already defined in
configure.ac, the benefit of having a ready to use spec file
outweighs the maintenance cost of the duplication.  The default values
can be overridden by passing corresponding --define command line
options to rpmbuild.

We cannot have a completely working autonomous vdsm.spec, because
package version and release are defined dynamically, by running
scripts from the source directory.  A spec file doesn't know the
location of the source directory when generating a SRPM and thus can't
run pkg-version script from it.  There are the following ways to have
a complete spec file:

- Running configure and make.

- Using build-aux/make-dist script (with or without running
  configure beforehand).

- Using vdsm.spec.in directly and defining vdsm_version and
  vdsm_release macros on rpmbuild command line, e.g.:

   rpmbuild -bs -D 'vdsm_version 4.40.0' -D 'vdsm_release 1' vdsm.spec.in

  The source tarball must be generated manually in this case.

Change-Id: I3ee44a76c8120dc6e8f17f75be62e25fdba1336d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>