net: Prepare QoS info report per southbound device
Currently, QoS info is included only at the network level.

In some scenarios, it is useful to detect the QoS configuration
on an individual device rather then on a (logical) network.
For example, when removing a network, the client (in our case a test)
may be interested to validate that the relevant QoS settings
have been removed from the underlying device.

This functionality will be used in a following change to
include the info in the network caps report.

Change-Id: Iac7dab638b8556309c64e761d87ad41d73d433da
Signed-off-by: Bell Levin <blevin@redhat.com>