docker: Add Fedora 30 dockerfile
This image contains only python 3 packages, and will be used to run vdsm
tests with python 3.7.

python3-ioprocess is not available yet on Fedora 30, so not included in
the image. Tests depending on ioprocess will fail until we add it.

With this image we can add a travis build for Fedora 30, running most of
the tests.

Change-Id: I46a3bb9a269bcf2fe0fb9da4ea57e103a7a2b42c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>