tests: Prepare for testing xlease with 4k storage
Current xlease tests create a new temporary file for every test, which
makes it hard to test with 4k storage or with a block device.

Refactor the test infrastructure so we use the same storage for all
tests. This will allow later to use user provided storage with 4k sector
size.

Add a tmp_storage fixture, that will be enhanced later with parameters
to run all tests with both 512 bytes and 4k sector size, and test
multiple alignments.

Add a tmp_vol fixture, using tmp_storage to format xleases index on the
provided storage.

The tmp_vol fixture returns a TemporaryVolume object, providing methods
to zero the storage and format the index, and writing records to the
index.

Change-Id: I8dc8edf9a926c69a7cd612cc74c83f1c701d2956
Bug-Url: https://bugzilla.redhat.com/1592916
Signed-off-by: Nir Soffer <nsoffer@redhat.com>